requirements.txt
streamlit
pandas
numpy
scikit-surprise
scikit-learn
joblib
tqdm
